requirements.txt
discord.py
discord-py-slash-command
discord-py-interactions